Revert "Move fee to frontend folder"
This reverts commit 2e867d931af9b2d64d93325858b2b36be67d1f94.
